desktop/requirements.txt
PySide6==6.6.0
httpx==0.25.2
python-dotenv==1.0.0
markdown==3.5.1
Pygments==2.17.2
requests==2.31.0
uvicorn[standard]==0.24.0
fastapi==0.104.1
python-multipart==0.0.6
pydantic==2.5.0
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rank-bm25==0.2.2
PyMuPDF==1.23.8
Pillow==10.1.0
numpy==1.24.3
aiofiles==23.2.1
playwright==1.40.0
pytesseract==0.3.10
opencv-python==4.8.1.78
unicodedata2==15.1.0